+++ .prettierrc.mjs
@@ -1,0 +1,10 @@
+/**
+ * @fileoverview Prettier configuration for @stackra/tsup-config
+ * @module @stackra/tsup-config
+ * @see https://prettier.io/docs/en/configuration
+ */
+
+/** @type {string} */
+const config = '@stackra/prettier-config';
+
+export default config;